Move language test functionality to a feature slice
