+++ lingetic-spring-backend/src/main/java/com/munetmo/lingetic/LanguageTestService/DTOs/Attempt/AttemptResponse.java
@@ -1,4 +1,4 @@
-package com.munetmo.lingetic.core.DTOs.Attempt;
+package com.munetmo.lingetic.LanguageTestService.DTOs.Attempt;
 
 public record AttemptResponse(String status, String comment, String answer) {
 }
